Separation and Addition
Split up app.py and formatted it to be more readable and accessible. Also added more information to home page.
