+++ app.py
@@ -9,49 +9,23 @@ import json
 from dotenv import find_dotenv, load_dotenv
 from flask_login import LoginManager, login_manager
 import flask
+from spoonacular import getrecipeswithcalories, usercalories
 from flask_sqlalchemy import SQLAlchemy
-import sqlalchemy
-from sqlalchemy.sql.expression import null
-from spoonacular import getrecipeswithcalories
 
 load_dotenv(find_dotenv())
-
 app = flask.Flask(__name__, static_folder="./build/static")
-bp = flask.Blueprint("bp", __name__, template_folder="./build")
 
-# Point SQLAlchemy to your Heroku database
+login_manager = LoginManager()
+login_manager.init_app(app)
+
 url = os.getenv("DATABASE_URL")
-
 if url and url.startswith("postgres://"):
     url = url.replace("postgres://", "postgresql://", 1)
 
 app.config["SQLALCHEMY_DATABASE_URI"] = url
-# Gets rid of a warning
 app.config["SQLALCHEMY_TRACK_MODIFICATIONS"] = False
-
 app.secret_key = os.urandom(9)
-
 db = SQLAlchemy(app)
-login_manager = LoginManager()
-login_manager.init_app(app)
-
-
-class CreateUser(db.Model):
-    """
-    Model for Users
-    """
-
-    id = sqlalchemy.Column(sqlalchemy.Integer, primary_key=True)
-    email = sqlalchemy.Column(sqlalchemy.String(120), unique=True)
-    name = sqlalchemy.Column(sqlalchemy.String(120), unique=True)
-    age = sqlalchemy.Column(sqlalchemy.String(3))
-    gender = sqlalchemy.Column(sqlalchemy.String(1))
-    weight = sqlalchemy.Column(sqlalchemy.String(3))
-    height = sqlalchemy.Column(sqlalchemy.String(3))
-
-
-# db.drop_all()
-db.create_all()
 
 
 @login_manager.user_loader
@@ -70,6 +44,7 @@ def catch_all(path):
     """
     return flask.render_template("index.html")
 
+bp = flask.Blueprint("bp", __name__, template_folder="./build")
 
 @bp.route("/index")
 # @login_required
@@ -86,6 +61,37 @@ def index():
 
 
 app.register_blueprint(bp)
+
+
+@app.route("/getuserinfo", methods=["POST"])
+def userinfo():
+    """
+    Send UserData if it exists to the frontend
+    """
+    try:
+        email = flask.request.json.get("email")
+        user = CreateUser.query.filter_by(email=email).first()
+        cal = usercalories(CreateUser.query.filter_by(email=user.email).first())
+        recipes = getrecipeswithcalories(cal)
+        data = {
+            "weight": user.weight,
+            "height": user.height,
+            "age": user.age,
+            "gender": user.gender,
+            "calories": cal,
+            "recipes": recipes,
+        }
+    except:
+        recipes = getrecipeswithcalories(2000)
+        data = {
+            "weight": 0,
+            "height": 0,
+            "age": 0,
+            "gender": 0,
+            "calories": 0,
+            "recipes": recipes,
+        }
+    return flask.jsonify({"data": data})
 
 
 @app.route("/login", methods=["POST"])
@@ -115,53 +121,11 @@ def login_post():
     return flask.jsonify({"loginResponse": "Ok"})
 
 
-@app.route("/getuserinfo", methods=["POST"])
-def userinfo():
-    """
-    Send UserData if it exists to the frontend
-    """
-    try:
-        email = flask.request.json.get("email")
-        user = CreateUser.query.filter_by(email=email).first()
-        cal = usercalories(user.email)
-        recipes = getrecipeswithcalories(cal)
-        data = {
-            "weight": user.weight,
-            "height": user.height,
-            "age": user.age,
-            "gender": user.gender,
-            "calories": cal,
-            "recipes": recipes,
-        }
-    except:
-        recipes = getrecipeswithcalories(2000)
-        data = {
-            "weight": 0,
-            "height": 0,
-            "age": 0,
-            "gender": 0,
-            "calories": 0,
-            "recipes": recipes,
-        }
-    return flask.jsonify({"data": data})
-
-
-def usercalories(useremail):
-    """
-    Get calories needed from the user
-    """
-    user = CreateUser.query.filter_by(email=useremail).first()
-    caloriesneeded = (
-        (10 * int(user.weight)) + (6.25 * int(user.height)) - (5 * int(user.age))
-    )
-    if user.gender == "F":
-        caloriesneeded -= 161
-    elif user.gender == "M":
-        caloriesneeded += 5
-    return caloriesneeded
-
-
 if __name__ == "__main__":
+    from database import CreateUser
+
+    db.drop_all()
+    db.create_all()
     # First app.run is local use. Second app.run is Heroku.
-    # app.run(use_reloader=True, debug=True)
+    #app.run(use_reloader=True, debug=True)
     app.run(host="0.0.0.0", port=int(os.getenv("PORT", 8080)))
